refactor: reorganize components by separating UI primitives from business logic
- Move RecentMatchesCard and StatsCard from ui/ to common/ directory
- Update import paths in Dashboard.tsx to reflect new structure
- Improve component architecture by separating pure UI components from business logic components
