+++ team-mars-frontend/src/components/common/recent-matches-card.tsx
@@ -179,5 +179,3 @@ export const RecentMatchesCard = () => {
         </Card>
     )
 }
-
-
