+++ team-mars-frontend/src/components/common/stats-card.tsx
